Making a nodata mask from the recorded polygons, skip loading these regions (#39)
* first pass at making a nodata mask from the recorded polygons

addressing #8

* move opera nodata poly to workflow._utils

* get rid of 3.11 restriction now that numba is okay

* move in old functions about eroding data

* test-build-push.yml: just use setuptools for version without building docker twice

* adjust parse/benchmark for first working version

* pass the config, not output log, to parse benchmark

* get new working nodata rasterize working

* track queued slices, remove redundant skipping

* implement mask_value data skipping for wrapped phaseD

* add version note to changelog

* move logging out of cli module into s1_disp

* change default docker image to ubuntu

* test-build-push: fix setup steps to split uses/run

* switch the ubuntu check to be an `apt-get` check

+++ scripts/benchmarking/benchmark-local.sh
@@ -5,14 +5,25 @@ set -o errexit
 set -o nounset
 set -o pipefail
 
+USAGE="Usage: $0 -b|--benchmark-dir [-i|--image <docker image>]"
+
 # Parse input arguments
 while [[ $# -gt 0 ]]; do
     key="$1"
     case $key in
-    -t | --tag)
-        TAG="$2"
+    -i | --image)
+        IMAGE="$2"
         shift
         shift
+        ;;
+    -b | --benchmark-dir)
+        BENCH_DIR="$2"
+        shift
+        shift
+        ;;
+    -h | --help)
+        echo "$USAGE"
+        exit 0
         ;;
     *)
         echo "Unknown option: $key"
@@ -21,17 +32,42 @@ while [[ $# -gt 0 ]]; do
     esac
 done
 
-if [ -z "${TAG+x}" ]; then
-    TAG="opera-adt/dolphin:latest"
-    echo "Using default tag: $TAG"
+if [ -z "${BENCH_DIR+x}" ]; then
+    echo "Please specify a benchmark directory."
+    echo "$USAGE"
+    exit 1
+fi
+
+if [ -z "${IMAGE+x}" ]; then
+    IMAGE="opera-adt/dolphin:latest"
+    echo "Using default docker image: $IMAGE"
 fi
 
 # Run the SAS workflow.
-BENCH_DIR="/u/aurora-r0/staniewi/dev/dolphin-benchmarks/"
 DATA_DIR="$BENCH_DIR/data"
-cd $DATA_DIR
 
-# WORKDIR="/tmp"
+# Check that the data directory exists.
+if [ ! -d "$DATA_DIR" ]; then
+    echo "Data directory does not exist: $DATA_DIR"
+    exit 1
+fi
+# Check we have input_slcs/ and dynamic_ancillary/ config_files/ and
+# config_files/dolphin_config.yaml
+if [ ! -d "$DATA_DIR/input_slcs" ]; then
+    echo "Input SLCs directory does not exist: $DATA_DIR/input_slcs"
+    exit 1
+fi
+if [ ! -d "$DATA_DIR/dynamic_ancillary" ]; then
+    echo "Dynamic ancillary directory does not exist: $DATA_DIR/dynamic_ancillary"
+    exit 1
+fi
+if [ ! -d "$DATA_DIR/config_files" ]; then
+    echo "Config files directory does not exist: $DATA_DIR/config_files"
+    exit 1
+fi
+
+cd $DATA_DIR
+WORKDIR="/tmp"
 mkdir -p scratch output
 rm -rf scratch/* output/*
 docker run --rm --user=$(id -u):$(id -g) \
@@ -41,16 +77,15 @@ docker run --rm --user=$(id -u):$(id -g) \
     --volume="$(realpath scratch):$WORKDIR/scratch" \
     --volume="$(realpath output):$WORKDIR/output" \
     --workdir="$WORKDIR" \
-    "$TAG" dolphin run dolphin_config.yaml
-
-dolphin run config_files/dolphin_config.yaml
-
-eval "$(conda shell.bash hook)"
-source ~/anaconda3/etc/profile.d/conda.sh
-conda activate mapping
+    "$IMAGE" dolphin run dolphin_config.yaml
 
 LOG_DIR="$BENCH_DIR/logs"
 mkdir -p $LOG_DIR
+
+# TODO: run this in a docker container.
+eval "$(conda shell.bash hook)"
+source ~/anaconda3/etc/profile.d/conda.sh
+conda activate mapping
 python /u/aurora-r0/staniewi/repos/dolphin/scripts/benchmarking/parse_benchmark_logs.py \
-    --config-files output/dolphin.log \
+    --config-files config_files/dolphin_config.yaml \
     --outfile "$LOG_DIR/benchmark_results_$(date +%Y%m%d).csv"
